Commit 12: minor final code refactoring
